Add  modular js
