+++ src/index.js
@@ -1,34 +1,5 @@
 import './style.css';
 import 'bootstrap/dist/css/bootstrap.min.css';
-import ToDo from './todo';
-import { addTodoLs, getTodosls, updateTodosLs } from './storage';
-import displayTodos from './display';
+import loadEventListeners from './load';
 
-const input = document.querySelector('#add-todo-input');
-const form = document.querySelector('#add-todo');
-
-form.addEventListener('submit', () => {
-  const description = input.value;
-  const newTodo = new ToDo(description);
-  addTodoLs(newTodo);
-});
-
-const toDosArr = getTodosls();
-displayTodos(toDosArr);
-
-const checkboxes = document.querySelectorAll('.check');
-
-checkboxes.forEach((check) => {
-  check.addEventListener('change', function () {
-    if (this.checked) {
-      const toDo = toDosArr[this.id];
-      ToDo.changeStatus(toDo);
-      updateTodosLs(toDo);
-    }
-    else if(!this.checked) {      
-      const toDo = toDosArr[this.id];
-      ToDo.changeStatus(toDo);
-      updateTodosLs(toDo);
-    }
-  });
-});
+loadEventListeners();
